Update following review + documentation
