+++ examples/Python/Basic/pointcloud.py
@@ -7,6 +7,21 @@
 import numpy as np
 from open3d import *
 
+def display_inlier_outlier(cloud,ind):
+    inlier_cloud = select_down_sample(cloud,ind)
+    outlier_cloud = select_down_sample(cloud,ind,True)
+
+    outlier_cloud.paint_uniform_color([0.1, 0.9, 0.1])
+    inlier_cloud.paint_uniform_color([0.9, 0.1, 0.1])
+    print("Showing outliers: ")
+    draw_geometries([outlier_cloud])
+
+    print("Showing inliers: ")
+    draw_geometries([inlier_cloud])
+
+    print("Showing both sets together: ")
+    draw_geometries([inlier_cloud,outlier_cloud])
+
 if __name__ == "__main__":
 
     print("Load a ply point cloud, print it, and render it")
@@ -40,3 +55,11 @@ if __name__ == "__main__":
     chair.paint_uniform_color([1, 0.706, 0])
     draw_geometries([chair])
     print("")
+
+    print("------ Statistical Oulier Removal -------")
+    cl,ind = statistical_outlier_removal(pcd,50,2)
+    display_inlier_outlier(pcd,ind)
+
+    print("------ Radius Oulier Removal -------")
+    cl,ind = radius_outlier_removal(pcd,60,0.05)
+    display_inlier_outlier(pcd,ind)
