+++ src/Python/Core/open3d_pointcloud.cpp
@@ -89,7 +89,7 @@ void pybind_pointcloud_methods(py::module &m)
             "extrinsic"_a = Eigen::Matrix4d::Identity());
     m.def("select_down_sample", &SelectDownSample,
             "Function to select points from input pointcloud into output pointcloud",
-            "input"_a, "indices"_a, "invert"_a = true);
+            "input"_a, "indices"_a, "invert"_a = false);
     m.def("voxel_down_sample", &VoxelDownSample,
             "Function to downsample input pointcloud into output pointcloud with a voxel",
             "input"_a, "voxel_size"_a);
@@ -100,10 +100,12 @@ void pybind_pointcloud_methods(py::module &m)
             "Function to crop input pointcloud into output pointcloud",
             "input"_a, "min_bound"_a, "max_bound"_a);
     m.def("radius_outlier_removal", &RemoveRadiusOutliers,
-            "Function to remove points that have less than nb_points in a given sphere of radius radius",
+            "Function to remove points that have less than nb_points"
+            " in a given sphere of a given radius",
             "input"_a, "nb_points"_a, "radius"_a);
     m.def("statistical_outlier_removal", &RemoveStatisticalOutliers,
-            "Function to remove points that are furtehr away from their neighbours in average",
+            "Function to remove points that are further away from their "
+            "neighbours in average",
             "input"_a, "nb_neighbours"_a, "std_ratio"_a);
     m.def("estimate_normals", &EstimateNormals,
             "Function to compute the normals of a point cloud",
